move displayTexts into separate file
